fixs

+++ alkos/kernel/include/modules/timing_constants.hpp
@@ -22,8 +22,8 @@ static_assert(kClockTicksInSecondSize == static_cast<size_t>(ClockType::kLastClo
 static constexpr Timezone kUtcTimezone = {
     .west_offset_minutes     = 0,
     .dst_time_offset_minutes = 0,
-    .dst_time_start_seconds  = -1,
-    .dst_time_end_seconds    = -1,
+    .dst_time_start_seconds  = static_cast<u16>(-1),
+    .dst_time_end_seconds    = static_cast<u16>(-1),
 };
 
 }  // namespace timing_constants
